a475d1-Build-Django-File-Sharing-Bundle-API/repository_after/requirements.txt
django>=5.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
